Modernize the library
This bulk of this work is moving most type parameters to being
associated types.  This eases integration with other image formats.

The ImageBuffer and Color types are removed, because they are no longer
needed.  Buffer2d is implemented for all types that implement
image::GenericImage, so we no longer need a wrapper, and because of
associated types, we can merge images that have different backing
structures, but share the same pixel type.