Allow packers to be stored inside structs
This patch simply uses a boxed dyn pointer to the actual packer
implementation. We could also use an enum with all our potential packers
but the memory overhead and the amount of times this function is called
doesn't warrant such optimization (yet).

Regardless, this allows users to just store a packer in any struct.

Fixes issue #56.